requirements.txt
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.3.0
requests==2.31.0
joblib==1.3.2
python-dotenv==1.0.0